Added more exit code checking, and support for -w and -h.

+++ ra.sh
@@ -19,7 +19,11 @@
 # lcreate -l -U
 # -c sha1
 
-PATH=/usr/bin:/bin; export PATH
+SERVER="-h _RADMIND_HOST"
+AUTHLEVEL="-w _RADMIND_AUTHLEVEL"
+EDITOR="vi"
+
+PATH=/usr/local/bin:/usr/bin:/bin; export PATH
 RETRY=10
 
 Yn() {
@@ -32,14 +36,20 @@ Yn() {
 }
 
 usage() {
-    echo "Usage:	$0 [-c] { trip | update | create | auto }" >&2
+    echo "Usage:	$0 [ -c | -h server | -w authlevel ] { trip | update | create | auto }" >&2
     exit 1
 }
 
-while getopts c opt; do
+while getopts ch:w: opt; do
     case $opt in
     c)	CHECKSUM="-csha1"
 	;;
+
+    h)	SERVER="-h $OPTARG"
+    	;;
+
+    w)	AUTHLEVEL="-w $OPTARG"
+    	;;
 
     *)   usage
 	;;
@@ -55,14 +65,30 @@ cd /
 
 case "$1" in
 update)
-    ktcheck -n -c sha1
-    if [ $? -eq 1 ]; then
-	Yn "Update command file and/or transcripts?"
-	if [ $? -eq 1 ]; then
-	    ktcheck -c sha1
-	fi
-    fi
+    ktcheck ${AUTHLEVEL} ${SERVER} -n -c sha1
+    case "$?" in
+    0)	;;
+
+    1)	Yn "Update command file and/or transcripts?"
+	    if [ $? -eq 1 ]; then
+		ktcheck ${AUTHLEVEL} ${SERVER} -c sha1
+		RC=$?
+		if [ $RC -ne 1 ]; then
+		    echo Nothing to update
+		    exit $RC
+		fi
+	    fi
+	;;
+
+    *) exit $?
+	;;
+    esac
+
     fsdiff -A -v ${CHECKSUM} -o /tmp/ra.fsdiff.$$ .
+    if [ $? -ne 0 ]; then
+	exit 1
+    fi
+
     if [ ! -s /tmp/ra.fsdiff.$$ ]; then
 	echo Nothing to apply.
 	rm /tmp/ra.fsdiff.$$
@@ -74,20 +100,39 @@ update)
     fi
     Yn "Apply difference transcript?"
     if [ $? -eq 1 ]; then
-	lapply ${CHECKSUM} /tmp/ra.fsdiff.$$
+	lapply ${AUTHLEVEL} ${SERVER} ${CHECKSUM} /tmp/ra.fsdiff.$$
+	case "$?" in
+	0)	;;
+
+	*)	exit $?
+		;;
+	esac
     fi
     rm /tmp/ra.fsdiff.$$
     ;;
 
 create)
-    ktcheck -n -c sha1
-    if [ $? -eq 1 ]; then
-	Yn "Update command file and/or transcripts?"
+    ktcheck ${AUTHLEVEL} ${SERVER} -n -c sha1
+    case "$?" in
+    0)	;;
+
+    1)	Yn "Update command file and/or transcripts?"
 	if [ $? -eq 1 ]; then
-	    ktcheck -c sha1
+	    ktcheck ${AUTHLEVEL} ${SERVER} -c sha1
+	    rc = $?
+	    if [ $rc -ne 0 ]; then
+		exit $rc
+	    fi
 	fi
-    fi
+	;;
+
+    *)	exit $?
+    	;;
+    esac
     fsdiff -C -v ${CHECKSUM} -o /tmp/ra.fsdiff.$$ .
+    if [ $? -ne 0 ]; then
+	exit 1;
+    fi
     if [ ! -s /tmp/ra.fsdiff.$$ ]; then
 	echo Nothing to create.
 	rm /tmp/ra.fsdiff.$$
@@ -100,12 +145,15 @@ create)
     Yn "Store difference transcript?"
     if [ $? -eq 1 ]; then
 	lcreate ${CHECKSUM} /tmp/ra.fsdiff.$$
+	if [ $? -ne 0 ]; then
+	    exit 1;
+	fi
     fi
     rm /tmp/ra.fsdiff.$$
     ;;
 
 trip)
-    ktcheck -qn -c sha1
+    ktcheck ${AUTHLEVEL} ${SERVER} -qn -c sha1
     case "$?" in
     0)
 	;;
@@ -113,12 +161,14 @@ trip)
 	echo Command file and/or transcripts are out of date.
 	;;
     *)
-	echo Command file and/or transcripts unknown.
+	exit $?
 	;;
     esac
 
     fsdiff -C ${CHECKSUM} -o /tmp/ra.fsdiff.$$ .
-    # Check return value?
+    if [ $? -ne 0 ]; then
+	exit 1
+    fi
     if [ -s /tmp/ra.fsdiff.$$ ]; then
 	echo Trip failure: `hostname`
 	cat /tmp/ra.fsdiff.$$
@@ -128,22 +178,27 @@ trip)
 
 auto)
     fsdiff -C ${CHECKSUM} -o /tmp/ra.fsdiff.$$ .
+    if [ $? -ne 0 ]; then
+	echo Auto failure: `hostname` fsdiff
+	exit 1
+    fi
     if [ -s /tmp/ra.fsdiff.$$ ]; then
-	echo Trip failure: `hostname`
+	echo Auto failure: `hostname` trip
 	cat /tmp/ra.fsdiff.$$
-	exit 0
+	exit 1
     fi
 
-    ktcheck -q -c sha1
+    # XXX - if this fails, do we loop, or justs report error?
+    ktcheck ${AUTHLEVEL} ${SERVER} -q -c sha1
     if [ $? -eq 1 ]; then
 	while true; do
 	    fsdiff -A ${CHECKSUM} -o /tmp/ra.fsdiff.$$
 	    if [ $? -ne 0 ]; then
-		echo fsdiff failed
+		echo Auto failure: `hostname`: fsdiff
 		exit 1
 	    fi
 	    if [ -s /tmp/ra.fsdiff.$$ ]; then
-		lapply -q ${CHECKSUM} /tmp/ra.fsdiff.$$ 2>&1 > /tmp/ra.lapply.$$
+		lapply ${AUTHLEVEL} ${SERVER} -q ${CHECKSUM} /tmp/ra.fsdiff.$$ 2>&1 > /tmp/ra.lapply.$$
 		case $? in
 		0)
 		    echo Auto update: `hostname`
@@ -163,7 +218,7 @@ auto)
 		    cat /tmp/ra.lapply.$$
 		    sleep ${RETRY}
 		    RETRY=${RETRY}0
-		    ktcheck -q -c sha1
+		    ktcheck ${AUTHLEVEL} ${SERVER} -q -c sha1
 		    ;;
 		esac
 	    fi
